requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
langchain
openai
streamlit
python-dotenv
